requirements.txt
#use this command to install the requirements:  python -m pip install -r requirements.txt

alembic==1.6.5
Faker==8.5.1
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
SQLAlchemy==1.4.17
